requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic>=2.9.0
pymongo==4.6.0
requests==2.31.0
email-validator==2.1.0
python-multipart==0.0.6
pypdf==3.17.4
openai==1.50.2
google-generativeai==0.8.2
numpy==1.26.4
scikit-learn==1.3.2
